requirements.txt
webdriver_manager
selenium
seleniumbase
flask
pillow